requirements.txt
affine==2.3.0
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
blessings==1.7
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.7
django-braces==1.14.0
django-cors-headers==3.2.1
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
earthengine-api==0.1.213
ee==0.2
future==0.18.2
# GDAL==2.4.4 # v2.4.4 good for all OS
geojson==2.5.0 # can remove in lean build
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
httplib2==0.18.0
httplib2shim==0.0.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
oauthlib==3.1.0
Pillow==7.0.0
protobuf==3.11.3
psycopg2==2.8.4 # comment out for MacOS install, use psycopg2-binary
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyproj==2.5.0
pyshp==2.1.0
python3-openid==3.1.0
pytz==2019.3
rasterio==1.1.3 # comment out for Windows 10
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Shapely==1.7.0 # comment out for Windows 10
six==1.14.0
snuggs==1.4.7
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
